fix: the firm floored every order twice and bought the cheap half

+++ src/shrap/risk_compliance/risk_officer/store.py
@@ -186,8 +186,8 @@ class DecisionRow:
     side: str | None
     approved: bool
     reason_code: str
-    requested_quantity: int | None
-    approved_quantity: int | None
+    requested_quantity: float | None
+    approved_quantity: float | None
     binding_limit: str | None
     strategy_ids: list[str]
     detail: dict[str, Any]
